Раскидал по папкам и всякого мутного намутил
Отделил реквест типы в отдельный файл
Отделил clck апи в отдельный файл
Удалил ts-node который тут даром не всрался
Перешёл на yarn
