+++ node_ts.json
@@ -1,17 +1,15 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
-  "display": "Node",
-  "_version": "18.0.0",
 
   "compilerOptions": {
-    "lib": ["es2019", "es2020.promise", "es2020.bigint", "es2020.string"],
-    "module": "Node16",
-    "target": "ES2019",
+    "lib": ["es2020"],
+    "module": "ES2022",
+    "target": "ES2022",
 
     "strict": true,
     "esModuleInterop": true,
     "skipLibCheck": true,
     "forceConsistentCasingInFileNames": true,
-    "moduleResolution": "Node16"
+    "moduleResolution": "NodeNext"
   }
 }
