+++ src/PreparedRequest/apiRequest.ts
@@ -1,5 +1,5 @@
-import Request from "../Request";
-import config from "../config";
+import Request from "../Network/Request";
+import config from "./config";
 
 export default function apiRequest() {
   return new Request(config.api);
