+++ src/PreparedRequest/config.ts
@@ -4,13 +4,13 @@ export default {
     port: 443,
     scheme: "https",
   },
-  clckApi: {
-    host: "clck.ru",
+  authApi: {
+    host: "oauth.yandex.ru",
     port: 443,
     scheme: "https",
   },
-  authApi: {
-    host: "oauth.yandex.ru",
+  clckApi: {
+    host: "clck.ru",
     port: 443,
     scheme: "https",
   },
