+++ src/index.ts
@@ -1,5 +1,5 @@
 import YMApi from "./YMApi";
 import WrappedYMApi from "./WrappedYMApi";
-import * as Types from "./types";
+import * as Types from "./Types";
 
 export { YMApi, WrappedYMApi, Types };
